Move files to appropriate directories
