+++ front-end/src/component/Delivery_app.js
@@ -1,7 +1,7 @@
 import {useState, useEffect} from "react";
 import {BrowserRouter, Switch, Route, Link} from "react-router-dom";
 
-import "./all.css";
+import "../all.css";
 
 let Delivery_app = (props) => {
 	let get_order_id = () => {
