+++ front-end/src/component/Header.js
@@ -1,7 +1,7 @@
 import {useState, useEffect} from "react";
 import {BrowserRouter, Switch, Route, Link} from "react-router-dom";
 
-import "./all.css";
+import "../all.css";
 
 let Header = (props) => {
 	return (
