+++ front-end/src/component/Menu_item.js
@@ -1,7 +1,7 @@
 import {useState, useEffect} from "react";
 import {BrowserRouter, Switch, Route, Link} from "react-router-dom";
 
-import config from "./config";
+import config from "../config";
 
 let Menu_item = (props) => {
 	let click_handler = async (ev) => {
