+++ front-end/src/component/Restaurant.js
@@ -1,7 +1,7 @@
 import {useState, useEffect} from "react";
 import {BrowserRouter, Switch, Route, Link} from "react-router-dom";
 
-import config from "./config";
+import config from "../config";
 
 let Restaurant = (props) => {
 	let generate_order_id = () => {
