+++ front-end/src/screen/Home_screen.js
@@ -2,8 +2,8 @@ import axios from "axios";
 import {useState, useEffect} from "react";
 import {BrowserRouter, Switch, Route, Link} from "react-router-dom";
 
-import config from "./config";
-import Restaurant from "./Restaurant";
+import config from "../config";
+import Restaurant from "../component/Restaurant";
 
 let Home_screen = (props) => {
 	//MAGIC 10-12 restaurants
